requirements.txt
fastapi
uvicorn
streamlit
requests
beautifulsoup4
selenium
python-dotenv